DesingBot-main/requirements.txt
pyTelegramBotAPI==4.27.0
flask==2.3.3
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0

pymupdf==1.23.0
pillow==10.0.0
numpy==1.24.3

scikit-learn==1.3.0
openai==1.3.0
sentence-transformers==2.2.2

typing-extensions==4.8.0